fix(cli): address additional PR review feedback for insights
- Fix negative balance formatting to show -$5.00 instead of $-5.00
- Wrap default case in block to fix Biome noSwitchDeclarations lint error
- Add .int() constraint to numDays schema to reject fractional values

+++ packages/cli-api-contract/src/index.ts
@@ -115,7 +115,7 @@ const dailyStatsSchema = z.object({
 const insightsContract = oc
   .input(
     z.object({
-      numDays: z.number().min(1).max(30).default(7),
+      numDays: z.number().int().min(1).max(30).default(7),
       timezone: z.string().default("UTC"),
     }),
   )
